requirements.txt
django==4.2.8
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
django-filter
